chore(cli): move manual GitHub helpers into dedicated module
Relocated the GitHub-specific manual helpers from src/cli/lib/manual/utils.js to src/cli/lib/manual/github.js so the generic utilities file only exposes shared cache/verbosity helpers while platform-specific logic sits under a focused module.
